add autofill of gas price parameters
